// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-workspace-delta-change-body.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface LinkingWorkspaceDeltaChangeBody {
  ayahId: number | null;
  ayahOverrideIds: Array<number> | null;
  descriptions: Array<string> | null;
  enabled: boolean | null;
  included: boolean | null;
  kind: string | null;
  label: string | null;
  mode: string | null;
  quranWordId: number | null;
  selected: boolean | null;
  shape: string | null;
}