// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-workspace-delta-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingWorkspaceDeltaChangeResponse } from '../models/linking-workspace-delta-change-response';
export interface LinkingWorkspaceDeltaResponse {
  linkingDataRevision: number;
  normalizedAppliedChanges: Array<LinkingWorkspaceDeltaChangeResponse>;
  sourceId: number;
  sourceVersion: number;
  workspaceVersion: number;
}